% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Continuous annuity.R
\name{caxnc}
\alias{caxnc}
\title{Continuous annuities.}
\usage{
caxnc(actuarialtable, x, n, i, m)
}
\arguments{
\item{actuarialtable}{An actuarial table object.}

\item{x}{Age of the annuitant.}

\item{n}{Number of terms of the annuity, if missing annuity is intended to be paid until death.}

\item{i}{Interest rate (default value the interest of the life table). (should be a scalar).}

\item{m}{Deferring period. Assumed to be 1 whether missing.}
}
\description{
This function calculates the actuarial value of continuous annuities, given an actuarial table. Temporary and deferred annuities can be evaluated.
}
