\name{clifecontingencies-package}
\alias{clifecontingencies-package}
\alias{clifecontingencies}
\docType{package}
\title{
Package to perform actuarial math on continuing life contingencies.
}
\description{
	The clifecontingencies package performs actuarial mathematics calculation. The main objective of the package is to provide tools to carry out continuous calculation of life insurance.
}
\details{
\tabular{ll}{
Package: \tab clifecontingencies\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2022-12-01\cr
License: \tab CC0\cr
LazyLoad: \tab true\cr
}

}
\author{
Pedro Henricky S. Diniz
Maintainer: <henricky.profissional@gmail.com>
}
\references{
          DICKSON, D. C.; HARDY, M. R.; WATERS, H. R. Actuarial mathematics for life
          contingent risks. 2. ed. United Kingdom: Cambridge University Press,2016.

          FERREIRA, Paulo Pereira. Matemática atuarial: risco de pessoas. Rio de Janeiro: ENS, 2019.

          PROMISLOW, S. D. Fundamentals of actuarial mathematics. [S.l.]: John Wiley & Sons, 2014.


                      }
\seealso{
	\code{\link{cAxn}}, \code{\link{cAxyzn}}
}
\section{Warning }{
	This package and functions herein are provided as is, without any guarantee regarding the	accuracy of calculations. The author disclaims any liability arising by any losses due to direct or indirect use of this package. }

